pyproject.toml
[project]
name = "gupo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.11.0",
    "bitsandbytes>=0.48.2",
    "datasets>=4.4.1",
    "deepspeed>=0.18.2",
    "hydra-core>=1.3.2",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "numpy>=1.26.4",
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "peft>=0.17.1",
    "python-dotenv>=1.2.1",
    "scipy>=1.16.3",
    "sentence-transformers>=5.1.2",
    "torch==2.5.1",
    "transformers==4.57.1",
    "trl>=0.25.0",
    "vllm>=0.7.3",
    "wandb>=0.22.3",
]
